api/requirements.txt
# Greenspot Grocer REST API Requirements
# Install with: pip install -r requirements.txt

# FastAPI framework and dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Authentication and security
python-jose[cryptography]>=3.3.0
cryptography>=3.4.8
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Database connectivity
mysql-connector-python>=8.0.0
SQLAlchemy>=2.0.0
alembic>=1.12.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0

# HTTP client for testing
httpx>=0.25.0
requests>=2.31.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Monitoring and logging
structlog>=23.0.0
prometheus-client>=0.17.0

# Validation
email-validator>=2.1.0
pydantic[email]>=2.4.0

# CORS support (included in FastAPI but explicit)
starlette>=0.27.0

# Production server
gunicorn>=21.2.0

# Environment management
python-dotenv>=1.0.0

# Type hints
typing-extensions>=4.8.0

# JSON handling
orjson>=3.9.0

# Date/time handling
python-dateutil>=2.8.0
pytz>=2023.3